backend/pdfToQuiz/requirements.txt
langchain==0.0.171
openai==0.27.4
openapi-schema-pydantic==1.2.4
pypdf==3.7.0
streamlit==1.21.0
fpdf==1.7.2
python-dotenv
flask_cors
flask
